Don't change repo if we've not set one

+++ inc/app-repository-remove.sh
@@ -1,17 +1,19 @@
 #!/bin/bash
 echo
-echo -e $YELLOW"--->Removing $APPTITLE repository..."$ENDCOLOR
-GREPOUT=$(grep ^ /etc/apt/sources.list /etc/apt/sources.list.d/* | grep $APPREPOSITORYNAME)
-if ! [ "$GREPOUT" == "" ]; then
-    if [ "$REPOPPA" = 'YES' ]; then
-        sudo add-apt-repository -r -y $APPREPOSITORYLINK
+if [ ! "$APPREPOSITORYLINK" == "" ]; then
+    echo -e $YELLOW"--->Removing $APPTITLE repository..."$ENDCOLOR
+    GREPOUT=$(grep ^ /etc/apt/sources.list /etc/apt/sources.list.d/* | grep $APPREPOSITORYNAME)
+    if ! [ "$GREPOUT" == "" ]; then
+        if [ "$REPOPPA" = 'YES' ]; then
+            sudo add-apt-repository -r -y $APPREPOSITORYLINK
+        else
+            sudo rm /etc/apt/sources.list.d/$APPNAME.list
+            echo 'OK'
+        fi
+
+        source $SCRIPTPATH/inc/app-keys-remove.sh
+
     else
-        sudo rm /etc/apt/sources.list.d/$APPNAME.list
-        echo 'OK'
+        echo $APPTITLE' repository doesnt exist'
     fi
-
-    source $SCRIPTPATH/inc/app-keys-remove.sh
-
-else
-    echo $APPTITLE' repository doesnt exist'
 fi
